Add files via upload
The mintclassicsDB.sql file that contains the script to create and populate the Mint Classics relational database. The script includes commands to create the schema (called "mintclassics"), along with the tables and fields, primary and foreign keys, and the data. When the script has run successfully you will have a working relational nine-table database populated with data from the Mint Classics company.
 For reference and use in analysis, the EER (Extended Entity-Relationship) diagram that models the structure of the Mint Classics database is provided.