Uploaded Project Analysis Files: SQL Queries and Output CSV Files
1. SQL Queries:
The SQL scripts included in this folder were designed to extract and analyze key information from the Mint Classics database, focusing on inventory management, sales trends and Warehouse Utilization. These queries address business-critical questions.
2. Output CSV Files:
The output files generated from the SQL queries are stored in this folder in CSV format for easy accessibility and further exploration. These files capture the result sets of the queries and serve as the basis for recommendations provided to the Mint Classics Company.